requirements.txt
fastapi
PyYAML
streamlit
whisperx
python-multipart
piper-tts
